practice-4/backend/pyproject.toml
[tool.poetry]
name = "practice-2"
version = "0.1.0"
description = ""
authors = ["Rustam Akimov <akirusprod@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
numpy = "<2.0.0"
transformers = "^4.38.1"
tokenizers = "^0.15.2"
termcolor = "^2.4.0"
chromadb-client = "^0.4.25.dev0"
fastapi = "^0.115.2"
uvicorn = "^0.27.1"
python-dotenv = "^1.0.1"
torch = "^2.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
